New widgets : Button
The core.coffee provides a simple jquery plugins that allow to replace inputs in a file by their corresponding widgets.
The index.html offer a demonstration of inputs replacement in a real world example.
Added the build file I use to produce the compiled js and the documentation.
Widgets now flags their target to prevent extra replacements.
ColorPicker now open the dialog with enter or space.
ColorPickerDialog now return the focus on the colorpicker on close.
The default ColorPicker instance is now added on document.ready.